Rework the MFMA API to not explicitly mention the instruction.
Now, per design plans, `amdgpu.mfma` is a function of m, n, k, and
Nblks, which will be used to select the correct intrinsic.

We also add attributes for negation (as supported by gfx940) and make
`blgp` explicitly an enum for clarity. In addition, there's now a
`reducePrecision` attribute that lets you pull in the xf32 operations.