Adds support for transformed fusion output
Currently, the AlignTiling had the assumption
that the operator being fused will have the same
shape as the gemm output buffer. However, this
can change due to view like ops.

This commit adds support for tranform the
gemm store coordinates to match the shape of
the output of the fused operator should there
be a difference.